+++ src/collective/recipe/template/genshitemplate.py
@@ -11,5 +11,5 @@ class Recipe(Base):
         context = Context(parts=self.buildout, options=self.options)
         try:
             self.result = template.generate(context).render()
-        except UndefinedError, e:
+        except UndefinedError as e:
             raise zc.buildout.UserError("Error in template %s:\n%s" % (self.input, e.msg))